requirements.txt
# 深度学习框架
torch==2.70
torchvision==0.22.0

# 数据处理
pandas==2.0.3
numpy==1.23.5

# 图像处理
Pillow==11.2.1

# 可视化
matplotlib==3.7.2
d2l==1.0.3

